feat: moving generators [IA-232]
